+++ Sources/MapboxSearchUI/Helpers/PlaceholderView.swift
@@ -8,14 +8,15 @@ class PlaceholderView: UIView {
 
     private func placeholdedView() -> UIView {
         guard let restorationIdentifier else {
-            print("PlaceholderView: RestorationIdentifier is empty. Cannot instantiate an replacement")
+            _Logger.searchSDK
+                .warning("PlaceholderView: RestorationIdentifier is empty. Cannot instantiate an replacement")
             return self
         }
 
         let nibName = restorationIdentifier.trimmingCharacters(in: .decimalDigits)
         let nib = UINib(nibName: nibName, bundle: .mapboxSearchUI)
         guard let view = nib.instantiate(withOwner: nil, options: nil).first as? UIView else {
-            print("PlaceholderView: First object in nib is not a UIView (nibName=\(nibName)")
+            _Logger.searchSDK.error("PlaceholderView: First object in nib is not a UIView (nibName=\(nibName)")
             return self
         }
         view.translatesAutoresizingMaskIntoConstraints = translatesAutoresizingMaskIntoConstraints